Cleaning up the relationship between Moodle status and Mahara status
The Moodle submission has two independent status parameters:
1. Its "lock" flag can be LOCKED or UNLOCKED
2. Its submission status can be:
 - DRAFT: This is the status they start in if you require students to
  hit "submit"
 - SUBMITTED: The status after the student submits it, or after they save
  if you make the "submit" button not required
 - REOPENED: If you turn on multiple attempts, then each time you grant a
  new attempt it creates a new submission in this status.

The Mahara view can either be locked (i.e. submitted) or not

The Mahara view should be locked if and only if the Moodle submission
has the LOCKED flag AND/OR the SUBMITTED status. Any time we change
one of these parameters, we need to check the other one before we change
the Mahara view's status.