Refactor unlocking settings and add upgrade checks.
To preserve initial functionality the setting for specifying whether pages
needs to be unlocked after grading have been added to assignment settings.
The teacher has option of not locking Page at all, lock forever (like in
the case when feedback plugin has not been installed and locking was
enabled) and unlock after grading or on marks release if marking worflow is
in place. The patch is upgrading "locking" setting in existing assignment
where feedback plugin was enabled, making it "unlock after grading". The
patch is checking that assignfeedback_mahara is uninstalled blocking
upgrade, so that assignfeedback_mahara will not be conflicking with
event-based unlocking.